% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flow_in_day.R
\name{add_flow_in_day}
\alias{add_flow_in_day}
\title{add_flow_in_day}
\usage{
add_flow_in_day(
  points_df,
  area_sqmi,
  source_flow_col = "Flow",
  new_col = "flow_in_day"
)
}
\arguments{
\item{points_df}{data.frame with a flow field with a name indicated by the
source_flow_col input}

\item{area_sqmi}{numeric. Drainage area in sq mi, which may be provided by
\code{hydrotools::WaterGageBase$load_sf_da()}}

\item{source_flow_col}{character, designating column name for Flow, default
"Flow"}

\item{new_col}{character designating new column for Flow in watershed inches,
default "flow_in_day"}
}
\value{
data.frame with added column of "new_col" with flow in watershed
inches
}
\description{
Appends a flow timeseries data frame with a field to represent flow in
watershed inches
}
\details{
Calculates conversion of cfs to watershed inches with \code{convert.flow()}
Adds column of watershed in. flow to df
}
