% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGWdouble.R
\name{agwo_hspf}
\alias{agwo_hspf}
\title{agwo_hspf}
\usage{
agwo_hspf(S, C, dthr)
}
\arguments{
\item{S}{numeric of length 1 representing AGWS - active ground water storage
in inches}

\item{C}{numeric of length 1 representing the AGWRC - Active groundwater
recession coefficient}

\item{dthr}{numeric of length 1 representing the model timestep in hours}
}
\value{
A numeric of length 1 that is this timestep's AGWO
}
\description{
Calculate active groundwater outflow using HSPF methodologies
}
\details{
Calculate active groundwater outflow using HSPF formula of AGWO = S
\itemize{
\item (1 - C^timestep). This is an alternate of \code{agws::single_forecast()}
designed to help test against HSPF output
}
}
