% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_recession.R
\name{analyze_recession}
\alias{analyze_recession}
\title{analyze_recession}
\usage{
analyze_recession(df, site_name = "", min_len = 14, max_len = Inf)
}
\arguments{
\item{df}{data.frame with a "RecessionDay" logical field that indicates if a
day is characterized by receeding flow. This is identified via
\code{flag_stable_baseflow()}}

\item{site_name}{Chararacter. Site name, passed to output for informational
purposes only.}

\item{min_len}{Numeric. Value for minimum number of consecutive RecessionDays
to be consider an event, default is 14}

\item{max_len}{Numeric. Value for maximum number of consecutive RecessionDays
to be consider an event, default is Inf}
}
\value{
a list with df as the modified input now with potential baseflow
groupIDs and summary with a data.frame including only recession events
}
\description{
Identifies potential baseflow events from flow_csv
}
\details{
Separates recession events and finds start and end date for each event.
Duration and days between are calculated. Events are categorized by group ID.
Stops with warning if no events are identified.
}
