% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_event_stats.R
\name{attach_event_stats}
\alias{attach_event_stats}
\title{attach_event_stats}
\usage{
attach_event_stats(analysis_data, r_lim = 0)
}
\arguments{
\item{analysis_data}{df with Date, Flow, AGWR, delta_AGWR as minimum present
columns.}

\item{r_lim}{numeric. Minimum R.squared limit, default 0}
}
\value{
df with site_no, Date, Flow, AGWR, delta_AGWR, Year, Month,
Day, season, GroupID, calc_AGWRC, R_squared
}
\description{
Add baseflow event statistics, including recession coefficient, to a data
frame of baseflow events calculated from \code{agws::analyze_recession()}
}
\details{
Calculates baseflow recession coefficients after filtering data for in range
data points via \code{summarize_event()} via a log-linear regression using
\code{summarize_event()} and adds these regression statistics to the input
data frame of recession events. Optionally filters for events with regressions
with coefficients of determination (R squared) greater than user input r_lim
}
