% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_event_stats.R
\name{bf_event_stats}
\alias{bf_event_stats}
\title{bf_event_stats}
\usage{
bf_event_stats(data, flow_col = "Flow", date_col = "Date")
}
\arguments{
\item{data}{df with Flow, Date columns}

\item{flow_col}{Set column name for flow_col, default "Flow"}

\item{date_col}{Set column name for date_col, default "Date"}
}
\value{
List with AGWRC and R.squared for all elements
}
\description{
Linear model of flow data
}
\details{
Creates a linear model between flow and date columns in df.
Pulls AGWRC and R.squared values
}
