% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_standardize_analysis.R
\name{bf_standardize_analysis_df}
\alias{bf_standardize_analysis_df}
\title{bf_standardize_analysis_df}
\usage{
bf_standardize_analysis_df(df, gage_id)
}
\arguments{
\item{df}{df from locationTrimStats.csv requiring GroupID, Date, Flow, AGWR, delta_AGWR, kept, met_alpha columns}

\item{gage_id}{char usgs gage number from commandArgs in function call}
}
\value{
df with columns: Date, GroupID, Flow, AGWR, delta_AGWR, AGWRC, kept, met_alpha
}
\description{
Standardized column names
}
\details{
Input df is checked for specific column names, required names are stated if missing.
Outputs new df with standardized naming convention for common variables.
}
