% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lookup.R
\name{build_lookup}
\alias{build_lookup}
\title{build_lookup}
\usage{
build_lookup(obs_flow, da_sqmi, m, b, n_steps = 25)
}
\arguments{
\item{obs_flow}{Numeric. Vector of flow values to use as lookup column for
AGWRC}

\item{da_sqmi}{Numeric. Value of the drainage area for the gage in square miles}

\item{m}{Numeric. slope value for the gage in the relationship of AGWRC ~
log(Q)}

\item{b}{Numeric. intercept value for the gage in the relationship of AGWRC ~
log(Q)}

\item{n_steps}{numeric. number of rows/steps required in the lookup table,
ex. 25 is 25 steps, a larger value is more precise but has a longer compute
time}
}
\value{
A data frame containing a lookup table based on AGWRC, m, b, and flow
values that can be used to predict variables when missing some but having
others
}
\description{
build_lookup
}
\details{
A function that takes a vector of flows and create a lookup table to
estimate AGWRC from flows using user provided log-linear regression
coeffiicents (often provided by \code{fit_agwrc_regression()})
}
