% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_event_stats.R
\name{calc_event_stats}
\alias{calc_event_stats}
\title{calc_event_stats}
\usage{
calc_event_stats(data, event_num, dAGWRmax, dAGWRmin)
}
\arguments{
\item{data}{df with Date, Flow, AGWR, delta_AGWR as minimum present columns.
From locationBF.csv (analysis_df)}

\item{event_num}{i in 1:max(data$GroupID) (summarize_event), can be set to one specific GroupID if desired}

\item{dAGWRmax}{num 1 + dAGWR_range, default dAGWR_range is 0.03 from function call script (summarize_event)}

\item{dAGWRmin}{num 1 - dAGWR_range, default dAGWR_range is 0.03 from function call script (summarize_event)}
}
\value{
new df with data df, AGWRC, and R.squared
}
\description{
Creates new df with AGWRC and group R.squared
}
\details{
Creates a df based on 3 conditions in SQL, being AGWR, dAGWRmax, dAGWRmin.
New df is combined with AGWRC and R.squared from bf_event_stats
}
