% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.flow.R
\name{convert.flow}
\alias{convert.flow}
\title{convert.flow}
\usage{
convert.flow(flow_col, area_sqmi)
}
\arguments{
\item{flow_col}{numeric vector with flow in cfs}

\item{area_sqmi}{numeric of length one with drainage area in sq mi}
}
\value{
numeric vector flow in watershed inches
}
\description{
Convert flow in cfs to watershed inches
}
\details{
Convert a vector of flows in cfs to watershed inches using drainage area
}
