% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamStats_auto_delineation.R
\name{find_FeatureCollection}
\alias{find_FeatureCollection}
\title{find_FeatureCollection}
\usage{
find_FeatureCollection(JSON)
}
\arguments{
\item{state}{Character. Abbreviation of state of USGS gage, e.g. 'VA'}

\item{longitude}{Numeric. longitude value of USGS gage}

\item{latitude}{Numeric. latitude value of USGS gage}

\item{UID}{Character. USGS monitoring_location_id for output df e.g. that
derived from \code{read_waterdata_monitoring_location()}}
}
\value{
A df with list columns of UID, POINT geometry and POLYGON geometry for
a USGS gage
}
\description{
Get relevant features from a feature collection or other json
}
\details{
Using USGS streamStats service API https://streamstats.usgs.gov/ss-delineate/docs#/
Find "FeatureCollection" in JSON file, typically about 5 objects in bedded
from input df
}
