% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_agwrc_regression.R
\name{fit_agwrc_regression}
\alias{fit_agwrc_regression}
\title{fit_agwrc_regression}
\usage{
fit_agwrc_regression(event_df)
}
\arguments{
\item{event_df}{data.frame with GroupID, median_flow, event_AGWRC}
}
\value{
lm object with m, b, m-value, p-value, and R.squared
}
\description{
Runs linear regression of log(Flow) and eventAGWRC
}
\details{
Runs linear regression of log(Flow) and eventAGWRC from output of
\code{agws::make_event_regression_df()} to provide insight if recession decay
coefficients are related to flow levels (e.g. groundwater storage levels)
}
