% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_cooks.R
\name{flag_cooks}
\alias{flag_cooks}
\title{flag_cooks}
\usage{
flag_cooks(model)
}
\arguments{
\item{model}{An R object, typically returned by an lm object. This is often
derived in step 06 of the AGWS workflow}
}
\value{
A list with two vectors. cooks_d, representing the cooks distance
calcualte metric for each observation in model and cooks_flagged, a logical
vector that is TRUE when cooks values are beyond the recommended threshold
}
\description{
Calculates cooks distance from a regression R object and
determines potential outliers
}
