% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_outliers_IQR.R
\name{flag_outliers_IQR}
\alias{flag_outliers_IQR}
\title{flag_outliers_IQR}
\usage{
flag_outliers_IQR(values)
}
\arguments{
\item{values}{numeric. A vector of numeric values from which the IQR will be
calculated.}
}
\value{
logical. A vector of logical values that are TRUE if the
corresponding elment in values is out of the 1.5 * IQR
}
\description{
flag_outliers_IQR
}
\details{
Calculates the Inner Quartile Range (IQR; 25-75\%) of an input vector and
returns a logical vector that is TRUE when values are out of the 1.5 * IQR
}
