% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_stable_baseflow.R
\name{flag_stable_baseflow}
\alias{flag_stable_baseflow}
\title{flag_stable_baseflow}
\usage{
flag_stable_baseflow(
  df,
  flow_col,
  AGWR_col = "AGWR",
  delta_col = "delta_AGWR",
  delta_thresh = 0.03,
  max_gap = 3
)
}
\arguments{
\item{df}{df with flow_col, AGWR, and delta_AGWR columns}

\item{flow_col}{flow_col with Flow values from df}

\item{AGWR_col}{AGWR_col is assigned AGWR column from df}

\item{delta_col}{delta_col is assigned delta_AGWR column from df}

\item{delta_thresh}{Num threshold value for change in delta_AGWR}

\item{max_gap}{Num value for acceptable gap of false values, default is 3}
}
\value{
Logical vector re-defining recession days column
}
\description{
Evaluates recession days based on flow parameters
}
\details{
Runs through all recession day events and calculates if flow for a given index position
is >= the mean of flow * 1.15. If >=, recession day for index position is assigned false value.
}
\examples{
#Sample data for flow_col, AGWR, and delta_AGWR
obs_flow <- c(580, 544, 508, 544, 508, 473, 900, 403)
AGWR <- c(0.8430233, 0.9379310, 0.9338235, 1.0708661, 0.9338235,
0.9311024, 0.9260042, 0.9200913)
delta_AGWR <- c(0.8871349, 1.1125803, 0.9956207, 1.1467543,
0.8720264, 0.9970860, 0.9945246, 0.9936146)
#Assigning data frame of flow_col, AGWR, and delta_AGWR to df
df <- data.frame(obs_flow, AGWR, delta_AGWR)
flow_col <- "obs_flow"
#Running flag_stable_baseflow with df and flow_col
df <- flag_stable_baseflow(df, df[[flow_col]])
}
