% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_slope.R
\name{get_basin_slope}
\alias{get_basin_slope}
\title{get_basin_slope}
\usage{
get_basin_slope(gage_id)
}
\arguments{
\item{gage_id}{Character. The gage ID of a USGS gage e.g. "01631000" for S.F.
Shenandoah River at Front Royal}
}
\value{
The average watershed slope in percent
}
\description{
get_basin_slope
}
\details{
Calculates the average watershed slope from nhdplustools of a given
USGS gage
}
