% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drainage_area_sqmi.R
\name{get_drainage_area_sqmi}
\alias{get_drainage_area_sqmi}
\title{get_drainage_area_sqmi}
\usage{
get_drainage_area_sqmi(gage_id)
}
\arguments{
\item{gage_id}{character usgs gage number from commandArgs in function call}
}
\value{
numeric, drainage area in sq mi
}
\description{
Site drainage area
}
\details{
Find the drainage area of site based on gageID using dataRetrieval and will
use the old or new APIs bsaed on user package version. Deprecated in favor of
\code{hydrotools::WaterGageBase$load_sf_da()}
}
