% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heteroscedasticity.R
\name{Heteroscedasticity}
\alias{Heteroscedasticity}
\alias{heteroscedasticity}
\title{Heteroscedasticity}
\usage{
heteroscedasticity(model)
}
\arguments{
\item{model}{an object of class "lm" representing a linear regression, often
from a series of summarized baseflow events as may be generated from
\code{fit_agwrc_regression()} in the agws worflow}
}
\value{
A dataframe with five fields: the White test p-value, the
Breusch-Pagan test p-value, the White test statistic, the Breusch-Pagan test
statistic, and a T/F interpretation of the p-value. If either of the two
tests has a p-value < 0.05, heteroscedasticity is flagged True as a concern.
If both p-values are > 0.05, heteroscedasticity is flagged False as a concern.
}
\description{
Heteroscedasticity
}
\details{
Determines if heteroscedasticity is present at a given gage by
running the Breusch-Pagan test based on the model summary from the baseflow
workflow e.g. bptest(model)
Runs the White Test, a more general test for heteroscedasticity that does
not assume a specific functional form, based on the model summary from the
baseflow workflow e.g. bptest(model, ~fitted(model) + I(fitted(model)^2))
}
