% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_event_regression_df.R
\name{make_event_regression_df}
\alias{make_event_regression_df}
\title{make_event_regression_df}
\usage{
make_event_regression_df(points_df, regression_flow_col = "Flow")
}
\arguments{
\item{points_df}{data.frame with fields: Date, GroupID, Flow, AGWR,
AGWRC, kept, met_alpha}

\item{regression_flow_col}{Character, designating field name for
regression_flow_col, default "Flow"}
}
\value{
data.frame one row per recession event with fields GroupID,
start_date, end_date, n_days, median_flow, event_AGWRC
}
\description{
Summarized a data.frame to get median flow of recession events with corresponding AGWRC
}
\details{
Summarizes a data.frame of recession events to calculate median flow during
the event and the corresponding recession coefficient. Only keeps non-flagged
data, with flags inidcated by logical values in the fields kept and met_alpha.
The input for this function is often provided by \code{agws::trim_event_mk()}
or \code{agws::attach_event_stats()}. The output from this function may be
passed to \code{agws::fit_agwrc_regression()} for further analysis
}
