% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_group_count.R
\name{monthly_group_count}
\alias{monthly_group_count}
\title{monthly_group_count}
\usage{
monthly_group_count(event_df, date_col = "start_date", group_col = "GroupID")
}
\arguments{
\item{event_df}{data.frame. A data frame with fields that have, at a
minimumm, group ID and start date. Often derived from step 06 of DEQ bf
workflow}

\item{date_col}{Character, the name of the field with event start dates}

\item{group_col}{Character, the name of the field with event group IDs}
}
\value{
df with columns for month, event_cnt, and gage_total
}
\description{
monthly_group_count
}
\details{
creates a df with county of total groups and count of groups per month
}
