% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lookup.R
\name{site_factors}
\alias{site_factors}
\title{site_factors}
\usage{
site_factors(da_sqmi, flow_vec, vec_for_reg = NULL, m, b)
}
\arguments{
\item{da_sqmi}{numeric}

\item{flow_vec}{numeric}

\item{vec_for_reg}{numeric}

\item{m}{numeric}

\item{b}{numeric}
}
\value{
A data frame...
}
\description{
site_factors
}
\details{
A function...
}
