% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_scq.R
\name{solve_agwrc_lookup}
\alias{solve_agwrc_lookup}
\alias{solve_agwrc_log}
\title{solve_agwrc_lookup}
\usage{
solve_agwrc_lookup(S, Ctable)

solve_agwrc_log(S, m, b)
}
\arguments{
\item{S}{Input Storage}

\item{Ctable}{table of Storage and C for lookup}

\item{m}{slope of equation}

\item{b}{intercept of equation}
}
\value{
C from table with stair-step

C from table with stair-step
}
\description{
Returns a C value from a table with S as lookup, stair-step

Returns a C value from log lin regression
}
\details{
For use to looup a C value based on S. Works best with a very fine grained table.

For use to calculate a C with parameters developed by HARP process.
}
