% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_scq.R
\name{step_scq}
\alias{step_scq}
\title{step_scq}
\usage{
step_scq(numts, Sinit, method, in2cfs, b = 0, m = 1, C = 0.99)
}
\arguments{
\item{numts}{number of timesteps to iterate}

\item{Sinit}{initial storage}

\item{method}{how to simulate, can be 'static', 'lookup', ''solver' or a 2 compartment object}

\item{in2cfs}{conversion factor inches to cfs}

\item{b}{intercept for C = mS + b equation}

\item{m}{intercept for C = mS + b equation}

\item{C}{static C value}
}
\value{
dataframe with simulation values for each timestep from 1 to numts
}
\description{
Rund a model with multiple AGWS calcuation options.
}
\details{
User can choose to run a lookup version, regression version or
two compartment with varying C and max S per compartment.
}
