% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamStats_auto_delineation.R
\name{streamStats_Delineation}
\alias{streamStats_Delineation}
\title{streamStats_Delineation}
\usage{
streamStats_Delineation(state, longitude, latitude, UID)
}
\arguments{
\item{state}{Character. Abbreviation of state of USGS gage, e.g. 'VA'}

\item{longitude}{Numeric vector. longitude value of each USGS gage}

\item{latitude}{Numeric vector. latitude value of each USGS gage}

\item{UID}{Character vector. USGS monitoring_location_id for output df
\code{read_waterdata_monitoring_location()}}
}
\value{
A df with list columns of UID, POINT geometry and POLYGON geometry for all input USGS gages
}
\description{
WKT text for POINT and POLYGON of multiple USGS gage
}
\details{
Using USGS streamStats service API https://streamstats.usgs.gov/ss-delineate/docs#/
provides WKT file of multiple USGS gage watershed geometries, including POINT and POLYGON columns
}
