% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamStats_auto_delineation.R
\name{streamStats_Delineation_single}
\alias{streamStats_Delineation_single}
\title{streamStats_Delineation_single}
\usage{
streamStats_Delineation_single(state = "VA", longitude, latitude, UID)
}
\arguments{
\item{state}{Character. Abbreviation of state of USGS gage, e.g. 'VA'}

\item{longitude}{Numeric. longitude value of USGS gage}

\item{latitude}{Numeric. latitude value of USGS gage}

\item{UID}{Character. USGS monitoring_location_id for output df e.g. that
derived from \code{read_waterdata_monitoring_location()}}
}
\value{
A df with list columns of UID, POINT geometry and POLYGON geometry for a USGS gage
}
\description{
WKT text for POINT and POLYGON of USGS gage
}
\details{
Using USGS streamStats service API https://streamstats.usgs.gov/ss-delineate/docs#/
provides WKT file of USGS gage watershed geometry, including POINT and POLYGON columns
}
