% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_event.R
\name{summarize_event}
\alias{summarize_event}
\title{summarize_event}
\usage{
summarize_event(analysis_data, event_number = 0, dAGWR_range = 0.03)
}
\arguments{
\item{analysis_data}{df with Date, Flow, AGWR, delta_AGWR as minimum present
columns often calculated by \code{agws::analyze_recession()}}

\item{event_number}{Numeric. The event to analyze. If 0, this function will
analyze all events via \code{for} loop}

\item{dAGWR_range}{numeric. The allowable range of the change in AGWR (the
ratio of today's and yesterday's flow). This value is added and subtracted
from 1.0 to set the range e.g. the default value of 0.03 allows dAGWR from
0.97 to 1.03}
}
\value{
data.frame with all columns of analysis_data and now appended with
fields of calc_AGWRC as the slope of the event log-linear regression of
log(Q) ~ Date and the corresponding R squared
}
\description{
Creates new df with analysis_data, AGWRC and R.squared
}
\details{
Uses \code{calc_event_stats()} and, ultimately, \code{bf_event_stats()} to
format event data to remove out-of-range data and calculate log-linear
regression of baseflow recession for one or all events in the user provided
data frame analysis_data
}
