% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_event_mk.R
\name{trim_event_mk}
\alias{trim_event_mk}
\title{trim_event_mk}
\usage{
trim_event_mk(event_df, alpha = 0.2, min_len_mk = 3, flag_len = 14)
}
\arguments{
\item{event_df}{df from locationstats.csv}

\item{alpha}{num variable for threshold of significance, default 0.20}

\item{min_len_mk}{num variable for minimum Mann Kendall points, default 3}

\item{flag_len}{num variable that flags events as short, default 14}
}
\value{
df with num columns of best and original p-value, trim event start and end, window length, and
logical columns for meeting flag_len, alpha threshold and kept column
}
\description{
Runs Mann Kendall Test
}
\details{
Creates df with outcome from MK test,
df includes best and original p-value, trim event start and end, window length,
indicator if flag_len is met, indicator if alpha is met, and kept column
}
