% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flow_in_day.R
\name{add_flow_in_day}
\alias{add_flow_in_day}
\title{add_flow_in_day}
\usage{
add_flow_in_day(
  points_df,
  area_sqmi,
  source_flow_col = "Flow",
  new_col = "flow_in_day"
)
}
\arguments{
\item{points_df}{df with columns: Date, GroupID, Flow, AGWR, delta_AGWR, AGWRC, kept, met_alpha}

\item{area_sqmi}{num variable with drainage area in sq mi}

\item{source_flow_col}{char string designating column name for Flow, default "Flow"}

\item{new_col}{char string designating new column for Flow in watershed in., default "flow_in_day"}
}
\value{
point_df df with added column of "flow_in_day"
}
\description{
Adds watershed in. column
}
\details{
Calculates conversion of cfs to watershed in. with convert.flow function
Adds column of watershed in. flow to df
}
