% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_month_season.R
\name{add_month_season}
\alias{add_month_season}
\title{add_month_season}
\usage{
add_month_season(df, date_col = "Date")
}
\arguments{
\item{df}{data.frame input with column named "Date" in format "YYYY-mm-dd" as
Date obj}

\item{date_col}{Character. The field in df containing dates.}
}
\value{
data.frame now appended with columns Month and Season contianing num
month and char season for length of "Date" column
}
\description{
Add month and meteorologic season to a data.frame
}
\details{
Add numerical month column in data.frame based on Date column, then add char
meteorologic season column when month number matches individual season
parameters
}
\examples{
\dontrun{
#'#get data from N F Shenandoah River via Strasburg, VA - USGS 01634000
library(hydrotools)
StrasGage <- hydrotools::WaterGageDaily$new(gage_id = "01634000")
StrasGage$gage_data <- StrasGage$gage_data  |>
 dplyr::rename(Date = time)
month_season_data <- add_month_season(StrasGage$gage_data)
}
}
