% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGWdouble.R
\name{AGWdouble}
\alias{AGWdouble}
\title{AGWdouble}
\value{
reference class of type AGWdouble.
}
\description{
Object for calculating flows from a 2-compartment AGWS/hsp style surficial aquifer
}
\details{
Provides all variables and equations to perform simulatio nand keep track of mass balance
}
\examples{
NA
}
\seealso{
NA
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{agws1}}{Initial Storage in AGW compartment 1}

    \item{\code{agws2}}{Initial Storage in AGW compartment 2}

    \item{\code{c1}}{recession coeffiecent compartment 1 (def=0.99)}

    \item{\code{c2}}{recession coefficient in compartment 2 (def=0.99)}

    \item{\code{agwsmax1}}{maximum storage compartment 1 (default=1.0)}

    \item{\code{agwsmax2}}{maximum storage in compartment 2 (default=1.0)}

    \item{\code{tmethod}}{transfer method from 1 to 2 ("all", "trans")}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-NA-initialize}{\code{NA$new()}}
    \item \href{#method-NA-solve_double_C}{\code{NA$solve_double_C()}}
    \item \href{#method-NA-eval}{\code{NA$eval()}}
    \item \href{#method-NA-show_state}{\code{NA$show_state()}}
    \item \href{#method-NA-log_state}{\code{NA$log_state()}}
    \item \href{#method-NA-plot}{\code{NA$plot()}}
    \item \href{#method-unknown-clone}{\code{unknown$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-NA-initialize}{}}}
\subsection{\code{NA$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{NA$new(
  agws1 = 0,
  agws2 = 0,
  c1 = 0.99,
  c2 = 0.99,
  agwsmax1 = 1,
  agwsmax2 = 1,
  tmethod = "all"
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{agws1}}{Initial Storage in AGW compartment 1}
      \item{\code{agws2}}{Initial Storage in AGW compartment 1}
      \item{\code{c1}}{recession coeffiecent compartment 1 (def=0.99)}
      \item{\code{c2}}{recession coefficient in compartment 2 (def=0.99)}
      \item{\code{agwsmax1}}{maximum storage compartment 1 (default=1.0)}
      \item{\code{agwsmax2}}{maximum storage in compartment 2 (default=1.0)}
      \item{\code{tmethod}}{transfer method from 1 to 2 ("all", "trans")}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-solve_double_C"></a>}}
\if{latex}{\out{\hypertarget{method-NA-solve_double_C}{}}}
\subsection{\code{NA$solve_double_C()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{NA$solve_double_C(
  agws1,
  agws2,
  c1,
  c2,
  agwsmax1,
  agwsmax2,
  agwin1 = 0,
  tmethod = "all",
  dthr = 24
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{agws1}}{Initial Storage in AGW compartment 1}
      \item{\code{agws2}}{Initial Storage in AGW compartment 1}
      \item{\code{c1}}{recession coeffiecent compartment 1 (def=0.99)}
      \item{\code{c2}}{recession coefficient in compartment 2 (def=0.99)}
      \item{\code{agwsmax1}}{maximum storage compartment 1 (default=1.0)}
      \item{\code{agwsmax2}}{maximum storage in compartment 2 (default=1.0)}
      \item{\code{tmethod}}{transfer method from 1 to 2 ("all", "trans")}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-eval"></a>}}
\if{latex}{\out{\hypertarget{method-NA-eval}{}}}
\subsection{\code{NA$eval()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{NA$eval()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-show_state"></a>}}
\if{latex}{\out{\hypertarget{method-NA-show_state}{}}}
\subsection{\code{NA$show_state()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{NA$show_state(format = "markdown")}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-log_state"></a>}}
\if{latex}{\out{\hypertarget{method-NA-log_state}{}}}
\subsection{\code{NA$log_state()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{NA$log_state(ts)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-plot"></a>}}
\if{latex}{\out{\hypertarget{method-NA-plot}{}}}
\subsection{\code{NA$plot()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{NA$plot(type = "cfq", warmup = 0, ylim = TRUE)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unknown-clone"></a>}}
\if{latex}{\out{\hypertarget{method-unknown-clone}{}}}
\subsection{\code{unknown$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{unknown$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
