% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_recession.R
\name{analyze_recession}
\alias{analyze_recession}
\title{analyze_recession}
\usage{
analyze_recession(df, site_name = "", min_len = 0, max_len = Inf)
}
\arguments{
\item{df}{Applicable df with RecessionDay column}

\item{site_name}{Char value with site name}

\item{min_len}{Num value for minimum number of consecutive RecessionDays to be consider an event, default is 0}

\item{max_len}{Num value for maximum number of consecutive RecessionDays to be consider an event, default is Inf}
}
\value{
Outputs summarized df including only recession events grouped by ID
}
\description{
Consolidates flow_csv data into recession events
}
\details{
Separates recession events and finds start and end date for each event.
Duration and days between are calculated. Events are categorized by group ID.
Output df only includes recession events, non-events are removed.
}
