% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_event_stats.R
\name{attach_event_stats}
\alias{attach_event_stats}
\title{attach_event_stats}
\usage{
attach_event_stats(analysis_data, r_lim = 0)
}
\arguments{
\item{analysis_data}{df with Date, Flow, AGWR, delta_AGWR as minimum present columns.
From locationBF.csv (analysis_df)}

\item{r_lim}{num variable setting R.squared limit, default 0}
}
\value{
df with site_no, Date, Flow, AGWR, delta_AGWR, Year, Month,
Day, season, GroupID, calc_AGWRC, R_squared (locationstats.csv)
}
\description{
Creates new binded df between input df and function df
}
\details{
Joins analysis_df (locationBF.csv) with function df (summarize_event), uses SQL query
to filter for equal GroupID and Date, then filters out R.squared values less than 0
}
