% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_event_stats.R
\name{bf_event_stats}
\alias{bf_event_stats}
\title{bf_event_stats}
\usage{
bf_event_stats(data, flow_col = "Flow", date_col = "Date")
}
\arguments{
\item{data}{data.frame with two fields identified by flow_col and date_col}

\item{flow_col}{Character. Field name for y variable in the regression log(y)
~ x. default "Flow"}

\item{date_col}{Character. Field name for x variable in the regression log(y)
~ x. Default "Date"}
}
\value{
List with AGWRC (slope) and R.squared for regression
}
\description{
Calculate a log-linear regression
}
\details{
Creates a linear model between flow and date columns in data such as
log(flow_col) = f(date_col) Returns the slope coefficient converted to linear
space and the corresponding coefficient of determination (R squared)
}
