% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_standardize_analysis.R
\name{bf_standardize_analysis_df}
\alias{bf_standardize_analysis_df}
\title{bf_standardize_analysis_df}
\usage{
bf_standardize_analysis_df(df, gage_id)
}
\arguments{
\item{df}{data.frame to check}

\item{gage_id}{chararacter. A site name to append to the data if not present.}
}
\value{
data.frame with the columns: Date, GroupID, Flow, AGWR, delta_AGWR,
AGWRC, kept, met_alpha now formatted as character, numeric, or logical as
appropriate
}
\description{
Cehck for missing analysis fields in data.frame
}
\details{
Input df is checked for specific column names \code{c("GroupID", "Date",
"Flow", "AGWR", "delta_AGWR", "kept", "met_alpha")}, required names are stated
if missing. Formats these variables with standardized classes by ensuring
expected numeric data is numeric, expectedlogical is logical, and expected
character is character
}
