% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lookup.R
\name{build_lookup}
\alias{build_lookup}
\title{build_lookup}
\usage{
build_lookup(df_clean, da_sqmi, m, b, n_steps = 25, vec_for_reg = NULL)
}
\arguments{
\item{df_clean}{data frame}

\item{da_sqmi}{numeric}

\item{m}{numeric}

\item{b}{numeric}

\item{n_steps}{numeric}

\item{vec_for_reg}{numeric}
}
\value{
A data frame...
}
\description{
build_lookup
}
\details{
A function...
}
