% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGWR_utils.R
\name{calc_AGWR}
\alias{calc_AGWR}
\title{calc_AGWR}
\usage{
calc_AGWR(x)
}
\arguments{
\item{x}{Numeric vector with chronological values}
}
\value{
Numeric vector with AGWR in all indices except 1
}
\description{
Calculation of AGWR / Ratio of Iterative timeseries values
}
\details{
Active groundwater recession coefficient is AGWR = Qt / Qt-1 This function
divides the the value in the time series by the value in the previous
day/index of the series
}
\examples{
\dontrun{
#Get data from N F Shenandoah River via Strasburg, VA - USGS 01634000
library(hydrotools)
StrasGage <- hydrotools::WaterGageDaily$new(gage_id = "01634000")
StrasGage$gage_data$AGWR <- calc_AGWR(StrasGage$gage_data$value)
}
}
