% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGWR_utils.R
\name{calc_delta_AGWR}
\alias{calc_delta_AGWR}
\title{calc_delta_AGWR}
\usage{
calc_delta_AGWR(x)
}
\arguments{
\item{x}{Num vector with chronological AGWR values}
}
\value{
Num vector with change in AGWR in all positions except 1 and 2
}
\description{
Wrapper of calc_AGWR to comply with legacy code
}
\details{
Calls calc_AGWRC(x)
}
\examples{
\dontrun{
#get data from N F Shenandoah River via Strasburg, VA - USGS 01634000
library(hydrotools)
StrasGage <- hydrotools::WaterGageDaily$new(gage_id = "01634000")
StrasGage$gage_data$AGWR <- calc_AGWR(StrasGage$gage_data$value)
StrasGage$gage_data$deltaAGWR <- calc_delta_AGWR(StrasGage$gage_data$AGWR)
}
}
