% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_event_stats.R
\name{calc_event_stats}
\alias{calc_event_stats}
\title{calc_event_stats}
\usage{
calc_event_stats(data, event_num, dAGWRmax, dAGWRmin)
}
\arguments{
\item{data}{df with Date, Flow, AGWR, delta_AGWR as minimum present columns.}

\item{event_num}{Numeric. The groupID of the recession event}

\item{dAGWRmax}{numeric. The upper limit for the change in AGWR (which is the
ratio of today's and yesterday's flow). Defaults to 1.03}

\item{dAGWRmin}{numeric. The lower limit for the change in AGWR (which is the
ratio of today's and yesterday's flow). Defaults to 0.97}
}
\value{
data.frame of fitlered data with the recession regression slope and r
squared
}
\description{
Remove out-of-range data and calculates log-linear regression of a recession
event
}
\details{
Removes out-of-range data by filtering the flow values in data for days in
which the AGWR and dAGWR are within user provided range. Then, this filtered
data is passed to \code{bf_event_stats()} to calculate the log-linear
regression
}
