% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.flow.R
\name{convert.flow}
\alias{convert.flow}
\title{convert.flow}
\usage{
convert.flow(flow_col, area_sqmi)
}
\arguments{
\item{flow_col}{num column with flow in cfs}

\item{area_sqmi}{num variable drainage area in sq mi}
}
\value{
num column with flow in watershed in.
}
\description{
Convert cfs to watershed in.
}
\details{
Convert cfs to watershed in.
}
