% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamStats_auto_delineation.R
\name{find_FeatureCollection}
\alias{find_FeatureCollection}
\title{find_FeatureCollection}
\usage{
find_FeatureCollection(JSON)
}
\arguments{
\item{JSON}{JSON list that may contain featureCollections as derived in
\code{streamStats_Delineation_single()}}
}
\value{
List. The relevant data from the list.
}
\description{
Get relevant features from a feature collection or other json
}
\details{
Using USGS streamStats service API https://streamstats.usgs.gov/ss-delineate/docs#/
Find "FeatureCollection" in JSON file, typically about 5 objects in bedded
from input df
}
