% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_agwrc_regression.R
\name{fit_agwrc_regression}
\alias{fit_agwrc_regression}
\title{fit_agwrc_regression}
\usage{
fit_agwrc_regression(event_df)
}
\arguments{
\item{event_df}{df with GroupID, start_date, end_date, n_days, median_flow, event_AGWRC}
}
\value{
model variable with m, b, m-value, p-value, and R.squared
}
\description{
Runs linear regression of logFlow and eventAGWRC
}
\details{
Runs linear regression of logFlow and eventAGWRC
}
