% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardForecast.R
\name{forwardForecast}
\alias{forwardForecast}
\title{forwardForecast}
\usage{
forwardForecast(Q0, days = 0:90, AGWRC, m, b)
}
\arguments{
\item{Q0}{numeric of length 1. initial flow on day 0}

\item{days}{numeric vector. Days where forcast should be calculated and will
use corresponding AGWRC if provided vector. Will use a variable calculated
AGWRC if AGWRC = "lm_variable"}

\item{AGWRC}{numeric or character. The decay coefficient to regress Q0 for
forecast. May be a single numeric to allow for a constant forecast or a
vector of numeric values to allow for a variable forecast but must be of
length days. Otherwise, may be "lm_constant" to claculate a constant value
from m and b or "lm_variable" to have a variable value}

\item{m}{numeric of length 1 that is the slope of a log-linear relationship
of Q and AGWRC}

\item{b}{numeric of length 1 that is the intercept of a log-linear
relationship of Q and AGWRC}
}
\value{
A data frame containing the days, forcast flows and AGWRCs used in the forecast
}
\description{
forwardForecast
}
\details{
A function to calculate a forcast days in the future based on Q0. Allows for
vector inputs of days and AGWRC. Alternatively, allows for calculation of AGWRC
via AGWRC = "lm_constant" or AGWRC = "lm_variable"
}
