% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_fill.R
\name{gap_fill}
\alias{gap_fill}
\title{gap_fill}
\usage{
gap_fill(flag_vec, max_gap = 5)
}
\arguments{
\item{flag_vec}{Logical vector input (often calculated from AGWR, deltaAGWR and a
deltaAGWR threshold)}

\item{max_gap}{Numeric. Acceptable gap, default is 5}
}
\value{
Logical vector defining recession days now with gaps filled
}
\description{
Fills logical vectors with TRUE within max_gap range
}
\details{
Fills all false values with a consecutive length less than acceptable gap with
true values if sequence of false values are between true values. Used in
\code{flag_stable_baseflow()}, for instance, to fill in recession events by
allowing minor excursions
}
\examples{
#Sample AGWR and deltaAGWR values, with threshold for change in deltaAGWR and max_gap
AGWR <- c(0.8430233, 0.9379310, 0.9338235, 1.0708661, 0.9338235, 0.9311024, 0.9260042, 0.9200913)
delta <- c(0.8871349, 1.1125803, 0.9956207, 1.1467543, 0.8720264, 0.9970860, 0.9945246, 0.9936146)
df <- data.frame(AGWR, delta)
delta_thresh = 0.03
max_gap = 3
#Creates logical vector of recession events
flag_vec <- abs(delta - 1.0) < delta_thresh & AGWR < 1.0
#Applies gap_fill function to locate missed recession events
df$RecessionDay <- gap_fill(flag_vec, max_gap)
}
