% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drainage_area_sqmi.R
\name{get_drainage_area_sqmi}
\alias{get_drainage_area_sqmi}
\title{get_drainage_area_sqmi}
\usage{
get_drainage_area_sqmi(gage_id)
}
\arguments{
\item{gage_id}{char usgs gage number from commandArgs in function call}
}
\value{
num variable with drainage area in sq mi
}
\description{
Site drainage area
}
\details{
Find the drainage area of site based on gageID using hydrotools
}
