% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_event_regression_df.R
\name{make_event_regression_df}
\alias{make_event_regression_df}
\title{make_event_regression_df}
\usage{
make_event_regression_df(points_df, regression_flow_col = "Flow")
}
\arguments{
\item{points_df}{df with columns, Date, GroupID, Flow, AGWR, delta_AGWR, AGWRC, kept, met_alpha}

\item{regression_flow_col}{char string designating column name for regression_flow_col, default "Flow"}
}
\value{
df with GroupID, start_date, end_date, n_days, median_flow, event_AGWRC
}
\description{
Consolidates df for baseflow regression
}
\details{
Converts previous df into more organized df with limited columns. Groups data by GroupID, finds start_date
and end_date for all events, calculates median flow for each event and event AGWRC
}
