% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardForecast.R
\name{RegressionAGWRC}
\alias{RegressionAGWRC}
\title{RegressionAGWRC}
\usage{
RegressionAGWRC(Flow, m, b)
}
\arguments{
\item{Flow}{numeric of length 1. Flow to calculate AGWRC from in log-linear
relationship}

\item{m}{numeric of length 1 that is the slope of a log-linear relationship
of Q and AGWRC}

\item{b}{numeric of length 1 that is the intercept of a log-linear
relationship of Q and AGWRC}
}
\value{
A list containing the forcast flows and AGWRCs used in the forecast
}
\description{
RegressionAGWRC
}
\details{
Calculate AGWRC based on a log linear relationship of flow, m,
and b
}
