% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardForecast.R
\name{regressionLimitAGWRC}
\alias{regressionLimitAGWRC}
\title{regressionLimitAGWRC}
\usage{
regressionLimitAGWRC(
  Flow,
  m,
  b,
  low_flow_limit = NULL,
  low_agwrc_limit = NULL,
  high_flow_limit = NULL,
  high_agwrc_limit = NULL
)
}
\arguments{
\item{Flow}{numeric of length 1. Flow to calculate AGWRC from in log-linear
relationship}

\item{m}{numeric of length 1 that is the slope of a log-linear relationship
of Q and AGWRC}

\item{b}{numeric of length 1 that is the intercept of a log-linear
relationship of Q and AGWRC}

\item{low_flow_limit}{A flow value below which the regression is not
applicable. At this limit, the low_agwrc_limit will be returned. If NULL,
this is not considered and instead the regression is used.}

\item{low_agwrc_limit}{If the input flow is below the low_flow_limit, this
function will return this value}

\item{high_flow_limit}{A flow value above which the regression is not
applicable. At this limit, the high_flow_limit will be returned. If NULL,
this is not considered and instead the regression is used.}

\item{high_agwrc_limit}{If the input flow is above the high_flow_limit, this
function will return this value}
}
\value{
An AGWR coefficient calculated from the log-linear regression for Flow
}
\description{
regressionLimitAGWRC
}
\details{
Calculate AGWRC based on a log linear relationship of flow, m,
and b e.g. AGWRC = m * log(Flow) + b
}
