% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardForecast.R
\name{single_forecast}
\alias{single_forecast}
\title{single_forecast}
\usage{
single_forecast(Q0, AGWRC, days)
}
\arguments{
\item{Q0}{numeric of length 1 for initial flow}

\item{AGWRC}{numeric scalar or vector for AGWRC}

\item{days}{numeric scalar or vector running calculation up to maximum value in vector}
}
\value{
numeric column containing forecast flows
}
\description{
single_forecast
}
\details{
Calculate flow prediction based on initial flow, AGWRC, and days
}
