% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardForecast.R
\name{single_forecast}
\alias{single_forecast}
\title{single_forecast}
\usage{
single_forecast(Q0, AGWRC, days)
}
\arguments{
\item{Q0}{numeric of length 1 for initial flow}

\item{AGWRC}{numeric of length 1 for AGWRC}

\item{days}{numeric scalar or vector running calculation up to maximum value in vector}
}
\value{
numeric vector containing forecasted flows
}
\description{
single_forecast
}
\details{
Calculate flow prediction based on initial flow, AGWRC, and days
using the exponential decay function of Q = Q0 * (AGWRC^days)
}
\examples{
single_forecast(Q0 = 100, AGWRC = 0.97, days = 0:10)
}
