% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_event.R
\name{summarize_event}
\alias{summarize_event}
\title{summarize_event}
\usage{
summarize_event(analysis_data, event_number = 0, dAGWR_range = 0.03)
}
\arguments{
\item{analysis_data}{df with Date, Flow, AGWR, delta_AGWR as minimum present columns.
From locationBF.csv (analysis_df)}

\item{event_number}{num variable assigned to eventnum, default 0}

\item{dAGWR_range}{num variable for calculating dAGWRmax and dAGWRmin, default 0.03}
}
\value{
df with site_no, Date, Flow, AGWR, delta_AGWR, Year, Month,
Day, season, GroupID, calc_AGWRC, R_squared
}
\description{
Creates new df with analysis_data, AGWRC and R.squared
}
\details{
Creates variables that are used in calc_event_stats to calculate AGWRC and R.squared
and filter for unwanted data. Organized method of creating new df with all desired data
}
